+++ resources/assets/js/components/enso/vueforms/VueForm.vue
@@ -107,7 +107,7 @@
                 v-if="data.actions.destroy"
                 :disabled="data.actions.destroy.forbidden"
                 :class="data.actions.destroy.button.class"
-                @click.prevent="showModal = true">
+                @click.prevent="modal = true">
                 <span>{{ __(data.actions.destroy.button.label) }}</span>
                 <span class="icon">
                     <fa :icon="data.actions.destroy.button.icon"></fa>
@@ -146,10 +146,10 @@
             <div class="is-clearfix"></div>
         </form>
         <modal v-if="data.actions.destroy"
-            :show="showModal"
+            :show="modal"
             :__="__"
             :message="data.actions.destroy.button.message"
-            @cancel="showModal = false"
+            @close="modal = false"
             @commit="destroy()">
         </modal>
     </div>
@@ -191,7 +191,7 @@ export default {
     data() {
         return {
             loading: false,
-            showModal: false,
+            modal: false,
             errors: new Errors(),
         };
     },
@@ -247,7 +247,7 @@ export default {
             }, { _params: this.params });
         },
         destroy() {
-            this.showModal = false;
+            this.modal = false;
             this.loading = true;
 
             axios.delete(this.data.actions.destroy.path).then(({ data }) => {